--- src/main/java/youth/week6/member/entity/Sex.java
@@ -1,5 +1,0 @@
-package youth.week6.member.entity;
-
-public enum Sex {
-    MAIL,FEMALE
-}
