+++ src/main/java/youth/week6/member/member/entity/Sex.java
@@ -1,0 +1,5 @@
+package youth.week6.member.member.entity;
+
+public enum Sex {
+    MAIL,FEMALE
+}
