+++ src/main/java/youth/week6/member/member/repository/MembersRepository.java
@@ -1,7 +1,7 @@
-package youth.week6.member.repository;
+package youth.week6.member.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import youth.week6.member.entity.Members;
+import youth.week6.member.member.entity.Members;
 
 public interface MembersRepository extends JpaRepository<Members, Long> {
