+++ src/main/java/youth/week6/member/organizer/repository/OrganizersRepository.java
@@ -1,7 +1,7 @@
-package youth.week6.organizer.repository;
+package youth.week6.member.organizer.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import youth.week6.organizer.entity.Organizers;
+import youth.week6.member.organizer.entity.Organizers;
 
 public interface OrganizersRepository extends JpaRepository<Organizers, Long> {
