+++ src/main/java/youth/week6/member/participant/repository/AllergensRepository.java
@@ -1,7 +1,7 @@
-package youth.week6.participant.repository;
+package youth.week6.member.participant.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import youth.week6.participant.entity.Allergens;
+import youth.week6.member.participant.entity.Allergens;
 
 public interface AllergensRepository extends JpaRepository<Allergens, Long> {
