+++ src/main/java/youth/week6/member/participant/repository/ParticipantsRepository.java
@@ -1,7 +1,7 @@
-package youth.week6.participant.repository;
+package youth.week6.member.participant.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import youth.week6.participant.entity.Participants;
+import youth.week6.member.participant.entity.Participants;
 
 public interface ParticipantsRepository extends JpaRepository<Participants, Long> {
